Fixed several issues that came out of the code review.
The target log directory was not being created.
The log cutoff was not working properly. It was basically breaking on every line if the time was after the cutoff. That will teach me to do work past midnight.
Added an error message in the chat box if the log directory could not be created.
Removed the resources.resx file because it was overkill for such a simple usage.